fix: resolve all compiler warnings - dead_code, clippy, FFI safety (Phase 12 P0)
- Eliminate 121 compiler warnings down to 0
- dead_code: prefix unused fields, allow on unintegrated modules
- clippy: fix unused imports, strip_prefix, &PathBuf→&Path, derivable Default, clone on Copy, manual_map, etc.
- FFI: add allow(improper_ctypes_definitions) for Rust-to-Rust plugin system
- Fix Path/PathBuf type mismatch bug in doc_gen.rs
- All tests pass, zero warnings remaining

+++ crates/vais-jit/src/types.rs
@@ -115,7 +115,7 @@ impl TypeMapper {
             ResolvedType::Unit => 0,
             ResolvedType::Vector { element, lanes } => {
                 // Vector size = element size * lane count
-                self.size_of(element) * (*lanes as u32)
+                self.size_of(element) * *lanes
             }
             _ => {
                 if self.pointer_type == types::I64 {
